requirements.txt
fastapi
uvicorn
python-multipart
torch>=1.10.0
torchvision>=0.11.0
rasterio>=1.2.0
shapely>=1.8.0
geopandas>=0.10.0
numpy>=1.21.0
tqdm
pyyaml
boto3
scipy>=1.7.0
earthengine-api>=0.1.300
python-dotenv>=0.19.0
opencv-python-headless>=4.5.0
langchain
langchain-openai
langchain-core
httpx
pydantic
